+++ tests/local/transformers/test_validfromto_transformer.py
@@ -9,8 +9,8 @@ from pyspark.sql.types import (
 from spetlrtools.testing import DataframeTestCase
 from spetlrtools.time import dt_utc
 
+from spetlr.etl.transformers.validfromto_transformer import ValidFromToTransformer
 from spetlr.spark import Spark
-from spetlr.transformers.validfromto_transformer import ValidFromToTransformer
 
 
 class ValidFromToTransformerTest(DataframeTestCase):
